NE-8969 Move executions widget backend to regular stage backend (#2589)
