+++ backend/routes/Executions.types.ts
@@ -1,0 +1,3 @@
+import type { ExecutionGraphNode } from '../handler/ExecutionsHandler.types';
+
+export type ExecutionGraphResponse = Omit<ExecutionGraphNode, 'labels'>;
